#49 fix names and tests

+++ src/integration-test/java/by.ladyka.club/ClubEventTicketPriceIT.java
@@ -78,7 +78,7 @@ public class ClubEventTicketPriceIT implements Replacer {
         ClubEventTicketPriceDTO clubEventTicketPriceDTO = new ClubEventTicketPriceDTO();
         clubEventTicketPriceDTO.setType(EventTicketPriceType.dance);
         clubEventTicketPriceDTO.setCost(BigDecimal.valueOf(250L));
-        clubEventTicketPriceDTO.setQuantity(100);
+        clubEventTicketPriceDTO.setQuantity(100L);
         clubEventTicketPriceDTO.setStartActiveTime(LocalDateTime.now().minusHours(1));
         clubEventTicketPriceDTO.setEndActiveTime(LocalDateTime.now().plusDays(10));
         clubEventTicketPriceDTO.setEventId(1L);
@@ -103,7 +103,7 @@ public class ClubEventTicketPriceIT implements Replacer {
         )
                 .andExpect(status().isOk())
                 .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
-                .andExpect(jsonPath("$.items[0].quantity", is(clubEventTicketPriceDTO.getQuantity())))
+                .andExpect(jsonPath("$.items[0].quantity", is(clubEventTicketPriceDTO.getQuantity().intValue())))
                 .andExpect(jsonPath("$.items[0].cost", is(clubEventTicketPriceDTO.getCost().intValue())))
                 .andExpect(jsonPath("$.items[0].type", is(clubEventTicketPriceDTO.getType().toString())))
                 .andExpect(jsonPath("$.items[0].eventId", is(clubEventTicketPriceDTO.getEventId().intValue())))
@@ -129,7 +129,7 @@ public class ClubEventTicketPriceIT implements Replacer {
         clubEventTicketPriceDTO_forUpdate.setEventId(1L);
         clubEventTicketPriceDTO_forUpdate.setType(EventTicketPriceType.table);
         clubEventTicketPriceDTO_forUpdate.setCost(BigDecimal.valueOf(600L));
-        clubEventTicketPriceDTO_forUpdate.setQuantity(400);
+        clubEventTicketPriceDTO_forUpdate.setQuantity(400L);
         clubEventTicketPriceDTO_forUpdate.setStartActiveTime(_YESTERDAY_);
         clubEventTicketPriceDTO_forUpdate.setEndActiveTime(LocalDateTime.now().minusHours(1));
 
@@ -152,7 +152,7 @@ public class ClubEventTicketPriceIT implements Replacer {
         )
                 .andExpect(status().isOk())
                 .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
-                .andExpect(jsonPath("$.items[0].quantity", is(clubEventTicketPriceDTO_forUpdate.getQuantity())))
+                .andExpect(jsonPath("$.items[0].quantity", is(clubEventTicketPriceDTO_forUpdate.getQuantity().intValue())))
                 .andExpect(jsonPath("$.items[0].type", is(clubEventTicketPriceDTO_forUpdate.getType().toString())))
                 .andExpect(jsonPath("$.items[0].cost", is(clubEventTicketPriceDTO_forUpdate.getCost().intValue())))
                 //TODO: resolve problem with serialized LocalDataTime format
@@ -178,7 +178,7 @@ public class ClubEventTicketPriceIT implements Replacer {
         clubEventTicketPriceDTO_forUpdate.setEventId(1L);
         clubEventTicketPriceDTO_forUpdate.setType(EventTicketPriceType.table);
         clubEventTicketPriceDTO_forUpdate.setCost(BigDecimal.valueOf(600L));
-        clubEventTicketPriceDTO_forUpdate.setQuantity(400);
+        clubEventTicketPriceDTO_forUpdate.setQuantity(400L);
         clubEventTicketPriceDTO_forUpdate.setStartActiveTime(_YESTERDAY_);
         clubEventTicketPriceDTO_forUpdate.setEndActiveTime(LocalDateTime.now().minusHours(1));
 
@@ -201,7 +201,7 @@ public class ClubEventTicketPriceIT implements Replacer {
         )
                 .andExpect(status().isOk())
                 .andExpect(content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON))
-                .andExpect(jsonPath("$.items[0].quantity", is(clubEventTicketPriceDTO_forUpdate.getQuantity())))
+                .andExpect(jsonPath("$.items[0].quantity", is(clubEventTicketPriceDTO_forUpdate.getQuantity().intValue())))
                 .andExpect(jsonPath("$.items[0].type", is(EventTicketPriceType.dance.toString())))
                 .andExpect(jsonPath("$.items[0].cost", is(550D)))
                 //TODO: resolve problem with serialized LocalDataTime format
